refactor: replace AppBarComponent with TurnIndicatorComponent and introduce PlayerScoreComponent
- Removed the AppBarComponent and replaced it with a simpler AppBar in the online game screen.
- Added TurnIndicatorComponent to dynamically display player turns with appropriate colors and messages.
- Created PlayerScoreComponent to encapsulate player score display logic, improving code modularity and readability.
- Updated import paths to reflect the new component structure.
