+++ lib/presentation/screens/tic_tac_toe/components/player_score_component.dart
@@ -3,11 +3,15 @@ import 'package:flutter_design_system/flutter_design_system.dart';
 import 'package:jogo_da_velha/domain/enums/player_enum.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class PlayerScore extends StatelessWidget {
+class PlayerScoreComponent extends StatelessWidget {
   final PlayerEnum player;
   final int score;
 
-  const PlayerScore({super.key, required this.player, required this.score});
+  const PlayerScoreComponent({
+    super.key,
+    required this.player,
+    required this.score,
+  });
 
   @override
   Widget build(BuildContext context) {
